requirements.txt
langgraph>=0.2.0
pydantic-ai[openrouter]>=0.0.14
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
kaggle>=1.5.0
